Updates
Force constructor call to be constructor call.

Add defines to handle deprecation in Node 10.  The Node 9 module
version definition is used as there doesn't appear to be one for Node
10 (at least not in all versions of Node).

The biggest change is to avoid the Nan::New wrapping a
v8::Externalized version of the wrapped object. This can expose the
object to GC after the callback is completed (i.e. during iteration).